tests: Properly wrap open()
open() is actually a varargs function, taking an optional 3rd parameter (mode).
We were only implementing the 3-argument version (and then ignoring the 3rd parameter),
which is UB that has now started to bite us with crashes on ppc64el under umockdev.

DRM only ever calls the 2-parameter version, so change MockDRM to provide the 2-parameter
version (and noisily die if this assumption becomes false in the future).