test/graphics-dummy: Guard against unwanted symbol export.
GoogleTest *very conveniently* uses at least one global with a non-trivial
destructor. Due to the vagaries of static linking, these globals are
exported in the symbol table unless you explicitly hide them.

This interacts badly with `dlopen()`. The object will get constructed during
the initial loading process, along with the other globals, and then will
be constructed *again* when the ELF-module-constructor machinery runs
to load the `dlopen`ed DSO.

Even better, during process teardown the DSO's ELF-destructor will run
the destructor for the global. *Then* the executable's ELF-destructor will
run the destroctor for the global *again*. This is rarely what you want,
and in *this* case it results in GoogleTest hitting an `abort()` when it
tries to destroy a pthread_key_t a second time.

Hiding every symbol we don't explicitly want to export solves this. The
main executable has a GoogleTest global, and graphics-dummy.so has a
GoogleTest global, but they're now at different addresses and so we
don't double destruct a single object.