tests/graphics-dummy: Work around musl RTLD_NEXT bug.
`graphics-dummy.so` links in our `open()` interposer, which checks whether
a specific `open` call should be intercepted and uses `dlsym(RTLD_NEXT, …)` to
call the real `open` otherwse.

Weirdly, with musl libc, `dlsym(RTLD_NEXT, "open")` fails to find the symbol,
and returns `NULL` *if* `open` is not exported as a global symbol from
`graphics-dummy.so`.

As far as I can tell, this is a musl bug, but it's fairly annoying to try and
debug anything on Alpine. Exporting an `open` symbol is harmless, and avoids
the bug, so I choose to be lazy here.