mir::SystemExecutor: Handle destroying Workers from their own callbacks.
If a Worker is not recycled back into the thread pool then the last reference
to it will be freed when the `std::function<void()> work` is destroyed after
execution in the `work_loop`. Detect this in `~Worker()` and `detach()` rather
than `join()`ing the thread (which would deadlock).

The thread won't touch anything but stack-local memory, and will then
terminate itself.