mir::Server: Use sigaction rather than signal
Documentation claims that `signal()` has unspecified behaviour in multithreaded
programs (such as, for example, anything that uses Mir). This is really a
theoretical concern, as it *seems* to work.

More significantly, signal handlers installed using `sigaction()` can have a
different signature to extract more information about the signal. I have no
idea what happens if code registers such a signal handler using `sigaction()`
and we try and extract it using `signal()`, but it's asking for trouble.

Furthermore, if *we* don't use `sigaction()` then the signal handlers we try
to chain to will definitely not get the extra information they're expecting.