mir::Server: Handle SIGBUS in subsequent run_mir() calls
`libwayland` needs to install a `SIGBUS` handler in order to transform out-of-bounds SHM buffer
access caused by clients lying about the size of the SHM segment they've associated with the
wl_shm_buffer into a protocol error rather than a server crash.

This gets done, once per process, at the first `wl_shm_buffer_begin_access` call.

However! Mir also wants to handle `SIGBUS` as a fatal signal so that we can run emergency
cleanup. This means that the *second* non-concurrent call to `run_mir()` in a process will
replace `libwayland`'s `SIGBUS` handler with Mir's, and `libwayland` will not reinstall it.

The total result is that, in the second and subsequent calls to `run_mir()`, client errors
in wl_shm_buffer usage will result in Mir crashing rather than the client being disconnected
with a protocol error.

Fix this by capturing the current `SIGBUS` handler at the end of `run_mir()`, and restoring
it (instead of our own) on subsequen `run_mir()` calls.