frontend_wayland: Add StdLayoutUPtr
We need to store some unique pointers in structs that libwayland
uses `offset_of` on, and so must be StandardLayoutType.

`std::unique_ptr` is not *guaranteed* to be a StandardLayoutType,
and, indeed, with the libstdc++ from GCC 11 Clang does *not* think
that `std::unique_ptr` is a StandardLayoutType.

(See [the relevant LLVM bug](https://github.com/llvm/llvm-project/issues/53021))