platform: Document Display::configure harder.
It seems that not everyone can psychically intuit the invariant that the `DisplayBuffer`s
provided by a `Display` must remain vaild at all times, and are only invalidated by a call
to `Display::configure`.

This should really go in some higher-level narrative documentation, but making the call's
documentation more explicit is a cheap way to make things better.