renderers/gl: Clear framebuffer to opaque black, not fully-transparent.
The fully-transparent `glClearColor` was introduced in 2015 in commit dabbc2bc1f
titled “merge mir” (it is not a merge commit).

Since the practical result of a fully-transparent framebuffer is expensive black
on all hardware platforms (at least until transparent displays become a thing),
switch the behaviour *back* to clearing the framebuffer to black.