CMake: Add variable to enable compiler quirks
We have a magnificent history of hitting compiler warning false-positives
and ad-hoc-ing them off. Let's formalise this in a list of compiler
quirks we can pass in.

Ideally quirks should be of the form `filename:warning` in case we want
to turn this into a fully automated system in future; that is left as
as an exercise for the reader.