WaylandConnector: Fix use-after-free
The libwayland global filter is called during wl_global_remove
(sensibly, as it needs to know whether or not the client knows
about the global that's being removed).

This means that the filter will be called from the destructor
of any extension object that publishes a global (ie: all of them).

Ensure it lives long enough to avoid UAF.