mtd::StubBufferAllocator: Test SHM buffers better.
This *was* using its own `wl_shm_buffer_{begin,end}_access` invocations, which
makes it rather hard to test that `WlShmBuffer`'s SIGBUS handling is correct!

Instead, get a mapping from the `Buffer`'s `ReadMappableBuffer` implementation;
that is how the rest of Mir is expected to access it, and so tests real-world
use better.