platform/buffer_from_wl_shm: Fix double-free.
Mostly there's going to be a single `ShmBufferSIGBUSHandler` instance over a
process' lifetime, but our tests bring up and tear down many `Server`s in a
process.

We were double-freeing the `previous_handler` in that case - once in
`~ShmBufferSIGBUSHandler`, and then once when `install_sigbus_handler()`
tries to delete the old handler loaded from `sigaction()`.