gbm-kms: Have Display probe check EGL support
Until we finish decoupling the DisplayPlatform and the RenderingPlatform, we still
rely on the DisplayPlatform creating an EGL context for the RenderingPlatform, and
the DisplayPlatform itself relies EGL support to import GBM buffers.

Mir will not bring up a display on DRM nodes that do not support HW EGL, even if
they properly support KMS, so devices which bring up an llvmpipe-using EGL context.