tests/ShmBacking: Improve invalid access tests.
So, `mmap` will provide a mapping *of pages*, which means its going to
be a multiple of `PAGE_SIZE`, even if the backing file is not
a-multiple-of-`PAGE_SIZE` big. It zero-pads the end up to the next page,
*and accesses to this zero-initialised region do not fault*.

Now, it turns out I picked `4000` as the backing size, which is close enough
to the `PAGE_SIZE` of `4096` on *most* kernels Ubuntu builds that doubling
it puts you in a separate page, so these tests pass on all systems with a 4K
`PAGE_SIZE`.

ppc64el has a 64K `PAGE_SIZE` 😠

Fix this by querying `sysconf(_SC_PAGE_SIZE)` and making our backing
file *exactly* one page in size. Combined with `mmap` being specified to map
on a page-aligned address (which is really the only thing it could do), that
guarantees that *any* access outside the valid range is going to hit a
different page, and hence fault.

Fixes: #2774