tests/DisplayServerMainLoopEvents: Fix arm64 CI.
The `DisplayServerMainLoopEvents` tests sends various signals to the process from
a thread spawned from the `init` parameter of `run_mir`.

This means that it can race and send and receive a signal before the server is
actually started.

This *should* be fine, because the tests wait for the signal to be handled
before continuing, and this *should* wait for the mainloop to dispatch the
handler, which *should* only happen once the server is actually running.

*Some* part of that chain of *should*s fails on the qemu arm64 CI environment.

Rework the tests to wait for the server to start up before sending any
signals.