CMake: Add WLCS include path to wlcs-integration target
The integration doesn't need to *link* to anything in WLCS, but it *does*
need access to the headers that describe the interface it needs to provide.

By default WLCS installs these into the default /usr/include system path,
so we don't notice that we're missing them, but users who install in non-default
paths will need this.