LogindConsoleServices: Un-work-around puzzling logind behaviour.
When logind processes a `ReleaseDevice` request, it sends a final `PauseDevice("gone")` after
completing the DBus call. When *we* receive a `PauseDevice("gone")` event for a device we
have an active `Device` handle for we call `ReleaseDevice`.

During startup, we call `AcquireDevice`, use that device in probing, then `ReleaseDevice` multiple
times before the selected platform ends up calling `AcquireDevice` to get the fd that will be used
for the life of the compositor.

This means we see a whole bunch of `PauseDevice("gone")` events during startup. We work around
them for DRM devices by just ignoring this type of event entirely, but we don't do that for
input devices.

Since input platform probing generally only needs to check the *first* candidate device, this means
that we have been `ReleaseDevice`ing the first input device before we actually try to give it to
libinput, preventing the first enumerated input device from working at startup.

There's also something of a race here, between the `PauseDevice` signal and the platform trying to
`AcquireDevice`. But! The device probing and initial platform setup is done *before* our mainloop
is running, so DBus events are only processed when something is pumping the glib mainloop,
and the synchronous DBus calls we make here pump the loop. So the first time the `PauseDevice`
DBus signal can be handled is during or after the platform's `AcquireDevice` call, by which
point there's an entry in the `acquired_devices` list and the `PauseDevice` handler will
then helpfully remove it and call `ReleaseDevice`. Urgh.

So, don't do any of this.

Stop ignoring `PauseDevice("gone")` for DRM devices - DRM devices *can* legitimately go away -
and stop calling `ReleaseDevice` from the gone handler.

Fixes: #2837