platforms/evdev: Handle libinput device rejections better.
When libinput fails to add a device via `libinput_path_add_device` it will already have
called the close-fd hook we provide. We should not try to remove it from the `FdStore`
ourselves; that will (at best) result in `FdStore` generating a warning, or, if the
fatal stars align *just right*, will result in us closing a newly-acquired FD to an
unrelated input device.