tests/x11: Test refusal to start at a higher level
The `mgx::Platform` constructor has a precondition that it gets passed a valid
`std::shared_ptr<X11Resources>`. The new `DisplayProvider` code makes use
of that precondition, so we can't easily test for it in the constructor anymore
and bail on failure.

But (a) we're allowed to misbehave on precondition failures, and (b) the
only way for the rest of Mir to get into the `mgx::Platform` constructor
is via `create_display_platform`, which already checks this condition
before constructing the `mgx::Platform`.

So move the test up there, too.