graphics/probe: Rework `probe_module` helper.
`probe_module` doesn't need to do any processing of the arguments to the `probe`
function, so we can choose instead to pass in a nullary functor with those
arguments already bound.

This will allow the arguments to `probe_display_platform` and `probe_rendering_platform`
to diverge without needing to edit the helper.