platform/graphics: Improve RenderingPlatform selection.
Passing in the set of `DisplayInterfaceProvider`s to the probe lets us discard
rendering platforms that cannot work earlier.

This *shouldn't* be necessary, but this helps avoid a problem: we need whatever
platform is chosen to provide `the_buffer_allocator()` to integrate with each
platform providing a `GLRenderingProvider` to the renderer. The infrastructure
for that is there, but not yet hooked up, and this lets us handle everything
that currently works without needed the full solution implemented now.