UBSan: Die noisily and immediately on UB.
By default UBSan will log-and-continue when UB is detected. This results in
log messages, but doesn't change the exit code of the program.

This is unhelpful for CI, where we want *failures* on UB.

Set UBSan to immediately kill the program whenever UB is detected,
so CI passes imply no-UB-detected.