ASan: Don't do leak checking
Sadly, we *know* we have benign leaks in hard-to-fix places. It would still
be valuable to fail CI if ASan detects anything *else* - out of bounds writes,
use-after-free, etc.

Disable the leak checker so we can mark ASan failures as CI-fatal.
Hopefully we can get get the leaks fixed or correctly-suppressed *some*time!