renderer::gl::OutputSurface: Add `release_current`
It *was* the case that `gl::OutputSurface`s were only ever accessed
on a single thread; our `MultiThreadedCompositor` makes that guarantee.
That meant that `OutputSurface` could be made current exactly once
and never needed to be released.

Unfortunately, with `BasicScreenShooter` we don't have that guarantee;
rendering is done on a threadpool, so unless we have a thread-local
`Renderer`, we need to be able to release the underlying EGL context
in order that it can be made current on another thread.