renderers::gl::Renderer: Hook up `make_current()`/`release_current()`
`BasicScreenShooter` wants be able to use the same `gl::Renderer` on
an arbitrary thread - not simultaneously, but not always on the same
thread.

That means that thread-local state - such as the current EGL context -
cannot be assumed to be constant across calls to `render()`.

Also allow users to explicitly release the current context, via
the pleasantly-already-existing-`suspend()`.