CPUAddressableDisplayProvider: Expose supported pixel formats.
`DisplayProviders` almost certainly have a limited set of pixel formats
they can output, and renderers need to care about that in order for
colours to be correct (or in order to not be garbled, in the case of BPP
differences).

Add a `supported_formats()` query, and accept a `DRMFormat` in the
`alloc_fb()` call.