platform/DMA-buf: Handle cross-GPU buffers
When we have multiple output cards with renderers on each then the
`BufferAllocator` that allocated a `Buffer` is not necessarily the
one that will be providing the `GLRenderingProvider`.

In that case we need to re-import the dma-buf on the GPU that is
actually trying to render from it.

Set this up with a `DMABufEGLProvider` class that provides
an `as_texture()` method and delegate to that from the
relevant `GLRenderingProvider`s.