graphics/probe: Tweak manually-specified display-platform selection.
It's acceptable for a display platform's `probe` to return `SupportedDevices` with `support_level == unsupported`.
This is filtered out in our normal probe path.

Fix the manually-specified path to also filter out any devices with `support_level == unsupported` if
the platform claims to support *any* device. If not, continue with the same "add everything" behaviour.