platforms/eglstream-kms: Fixup.
We need to use the same EGLDisplay for the DisplayPlatform and the RenderingPlatform
(so they can share `EGLStreamKHR` objects), so plumb that through the platform probing.

Make sure there's a current EGLContext when constructing the EGLStreamOutputSurface
(thanks, libepoxy 😠).

Add some debugging for when DisplayBuffer::post() fails.

Ensure we've got up-to-date KMSConnector information when we go searching for CRTCs.