platforms/eglstream-kms: Just the worst bug.
So, it turns out that if your `EGLStreamProducerSurfaceKHR` is constructed in an EGL context
with an EGL config with non-zero alpha channel then everything will *appear* to work until,
when actually trying to get something on the screen, `eglStreamConsumerAquireAttribNV` will
fail with an undocumented `EGL_BAD_STATE_KHR` error.

Request a 0-sized alpha channel in `EGLStreamOutputSurface`.