platform: Split DisplayTarget into DisplayProvider & DisplayAllocator
One of the reasons that `DisplayTarget` is hard to name is that it is really doing
two separate (but related) jobs - there's (mostly) some display-global state that
needs to be shared, and then the ability to allocate something, (mostly) tied to
a specific `DisplayBuffer`.

Make those two things separate.

Now, rather than the `DisplayPlatform` -> `DisplayTarget` -> `DisplayProvider` we
have `DisplayPlatform` -> `DisplayProvider` for the display-wide interface.

Instead of `DisplayBuffer` -> `DisplayTarget` -> `DisplayProvider` we have
`DisplayBuffer` -> `DisplayAllocator` for the per-`DisplayBuffer` allocation
interface.

This also lets us hang the `DisplayAllocator` off the lifetime of the
`DisplayBuffer`; like a lot of the rest of the code, the `DisplayAllocator`
can now be valid only as long as the `DisplayBuffer` it's taken from.