graphics/probe: Implement new probe policy
The difference between Display and Rendering platform selection is that
Display prefers nested platforms if available, and Rendering prefers
hardware platforms if available.

Pass that policy selection into `modules_for_device`, and use it to
implement the new platform selection policy.

(Mostly; the Virtual platform needs to be treated separately, and that's
not done here)