MultiThreadedCompositor: Fix synchronisation of composition triggering.
Sadly, even though it *looks* like it should work, `std::binary_semaphore`
is just the wrong tool here.

Fortunately, C++20 *also* added “wait for this `std::atomic<>` value to change”,
which is more or less what we want. (It still has the annoying “you have to explicitly
`notify_one()`” bit which is not required by the underlying OS API, but 🤷‍♀️)