platform/DRMFormat: Don't require a `FormatInfo` for each format.
We're now using `DRMFormat` in a situation (linux-dmabuf) where:
* We don't actually need to care about many the properties of the format, and
* We're processing formats submitted by cilents.

This means that the existing “throw an exception if we haven't got a `FormatInfo`”
approach works badly; we will disconnect clients with a Wayland internal error
if they submit buffers in a format we haven't explicitly listed, even though it
*would* work if we just treated the format as an opaque identifier.

So, don't do that anymore. Instead, store the fourcc code directly in `DRMFormat`
alongside the `FormatInfo` if it exists.

Fixes: #3462