mo::DefaultConfiguration: Don't accidentally consume an unparsed option per module.
`mo::ProgramOption::parse_arguments()` assumes that the *first* argument in the arguments array is
`argv[0]`; that is, the name of the calling binary, and so it strips it out. This was causing us
to consume an unparsed argument for each module we had loaded.

Instead, insert `argv[0]` into the argument array before parsing.