platforms/atomic-kms: Fix `AtomicKMSOutput::set_power_mode`
We might not actually have a `current_crtc` when `set_power_mode` is called
(notably, when disabling an output we set `mir_power_mode_off`).

Rather than crashing, log an error if we try and turn on an unconfigured
output. Silently ignore trying to turn *off* an unconfigured output,
as it's already off.