platforms/atomic-kms: Fix modesetting with `AtomicKMSOutput::set_crtc`
The way changing the display mode works is that the
`configure` sets the desired `mode_index`, and then
the *next* `set_crtc` is meant to actually set the mode.

This means that we can wait for correctly-sized content
to present at the new mode, rather than showing a single
black frame before the correct content.

But it also means that `set_crtc` needs to look at
the requested mode, rather than the current mode!