platforms/atomic-kms: Don't try to pageflip incorrectly sized FBs
Currently we expect to only flip framebuffers that are
the same size, in pixels, as the output mode.

Check in `schedule_page_flip` that this holds. Otherwise,
we must be expecting a modeset and need to go through
the `set_crtc` codepath.

We might, later, update the system to handle using the
display scalers to present framebuffers that aren't the
same size as the output mode.