platforms/atomic-kms: Synchronise access to CRTC configuration.
*Almost* all of this is only accessed on a single thread, the compositor's
submission thread.

*Almost*.

Unfortunately, `configure` is generally called from the ServerAction loop,
which is a different thread to the composition thread. `ensure_crtc` can
also (transitively) be called from `configuration`, which also happens
off-composition-thread.

Wrap this ball of wax up in a `mir::Synchronised<>`, to ensure we're
not racy.

(This also affects `gbm-kms`, but to a much lesser extent)