platforms/atomic-kms: Remove wait_for_page_flip()
We now always have a single output per `DisplaySink`, so
we always immediately wait for the page flip to occur,
so there's no point in going through the dance of submitting
a non-blocking page filp and then immediately waiting for the
completion.

Instead, ask KMS to block until the page flip has occurred.