MIRENG-742: Fix stuck frame after mode switch.
When we perform a mode switch Mir first tears down all the `Compositor`s,
does the switch, and then creates a new set of `Compositor`s, one for each
output.

But, in the buffer queue code, we track which `Compositor`s have accessed
the “current” buffer of a surface, and only advance to the most recently
submitted buffer when the `Compositor` had already seen the current one.

After the mode switch, *all* the `Compositor`s are new, so *none* of them
will have seen the “current” buffer, and so will not receive the most
recently submitted client buffer until something else triggers a new
composite pass.

The buffer queue code is going to need to change significantly in the
forseeable future to accomodate the requirements of some Wayland extensions,
so work around this problem by having each `Compositor` thread mark
itself as a user of the “current” buffer on construction, so the
first composition pass will receive the most recent buffer.