Prevent `pending_state` from being reset before usage.
For some reason, we get a `reset` request between the state being set
and it being used whenever we tab between the interval (number) and the
text dropdown field to its right. This caused the pending state to be
reset, which would stop the state from being comitted to the text hub
handler, keeping the old state (number OSK) on a new field (text
dropdown).