platforms/atomic-kms: Fix cursors across multiple monitors.
`KMSOutput::has_cursor` was misleadingly named, and so we implemented it
incorrectly. It is *used* to tell whether or not a cursor image needs
to be set on that particular output before it is visible, *not* whether
that output *supports* a cursor.

Rename it to something hopefully more easily understandable, and
then fix the implementation.

Fixes: #3677