mg::gl::Texture: Add `tex_id` accessor.
It would be convenient for QtMir (and potentially other code) to
have access to the texture ID for the `gl::Texture`, and this is
(mostly) trivial for the implementations to provide, so do it.

The exception here is `ShmBuffer`, which can be created on a thread
with no current EGL context¹, and so can't do its GL setup on
creation. Instead, it currently defers that until `bind()` time,
where we know we have a current context.

We *do* have access to an EGL context in ShmBuffer, though,
through the `EGLContextExecutor` (though this may be a different
thread). So, instead queue the GL initialisation work in
the constructor, and have the texture ID resolve to a
`shared_future<GLuint>`.

¹: Something that has bitten us in the past, and we
might want to do something about, if we could just work out
*what* ☺  https://github.com/canonical/mir/discussions/2909