tests/ShmBuffer: Don't unnecessarily enforce cardinality in test assumptions.
Now that `ShmBuffer` *always* allocates a texture id, this test
which asserts that `gl{Gen,Delete}Textures` is called exactly once
becomes racy.

There's no need to enforce the exactly-once here; the test is testing
that the texture is deallocated on a thread with an EGL context current,
not how many times `glGenTextures` is called.