Fix deadlock in `BasicAccessibilityManager::notify_helpers`
The deadlock was caused by the for loop extending the lifetime of the
state lock to cover the duration of the whole loop. In the loop, we also
locked the state to read some data, which would cause the deadlock.

Why did this decide to come up now? I have no idea, it used to work
before (and should've failed in CI), but now it deadlocks...