platforms: Use `NativeBufferBase` in `as_texture`
Users may wish to wrap a `Buffer` in a class with some
extra behaviour of their own (notably, tracking when
the buffer is no longer used by the rendering pipeline).

In order to allow this wrapping while still having
`as_texture` access the *real* buffer type we should
be using `NativeBufferBase` so the wrapper can return
the underlying buffer.