SoftwareCursor: Capture weak references in `hide()` workitem
During server teardown we deliberately leak work items from the
`GLibMainLoop`; if those work items are capturing `shared_ptr`s
then we leak the whole subgraph.

`Cursor::hide()` can be called as a part of teardown (when the
pointer device is removed from the `Seat`), and so this work
item should only capture weak references.