platforms/gbm-kms: Enable DRM Universal Planes support. (#3978)
This is required for pulling the available output colour formats out of
KMS, as was enabled in #3950 (oops!).

`DRM_CLIENT_CAP_UNIVERSAL_PLANES` was unconditionally enabled in the
3.16 kernel; that's old enough for us to unconditionally enable it, too.

Fixes the crash on startup in `mgk::find_crtc_with_primary_plane`.
(Primary planes *only exist* when this client cap is set).