Expose the cursor's renerable via Cursor::renderable and access it in the compositor itself instead of the scene (#3980)
- Exposed `Cursor::renderable()` that will return the renderable
associated with the cursor, if there is any
- No longer adding the `Cursor::renderable()` to the `Scene`. Instead,
adding directly in the compositor
- Updated the associated `SoftwareCursor` implementation and tests for
this new interface
- Existing hardware implementations return `nullptr` for now (but I will
address this in later work, as I state below!)

The `Cursor` is not logically a part of the `Scene` as it is not a
surface or anything. Hence, it is up to the compositor to decide whether
or not it will display it.

This work isn't done, but this is the minimal reviewable chunk that I
want to get in before tackling the rest of it. Coming next:

1. Add `Cursor::needs_compositing` so that the compositor can rely on
that boolean flag to decide whether or not to add that `Cursor`'s
renderable
2. Implement `Cursor::renderable()` for each hardware cursor
3. Render the cursor in the `BasicScreenShooter` based off whether or
not an `include_cursor` flag is set