Construct the cursor multiplexer _after_ the main loop it depends on.
So what I think was happening is that the main loop would get
constructed, we got a reference to the `Executor` part of it, then it
would get destroyed because we only held a reference and not a shared
pointer. Then when we tried using the executor later on, we'd use the
invalid reference to try and spawn some work.

My assumption was that the main loop was one of the things to be created
in the server. Which apparently doesn't hold for
`CursorObserverIntegrationTest`.

Alan was right once again!