Remove --enable-input
This option was added in 2013 in f864944fc3de3b6a4953c90189a503349f0e60a5 and defaulted to false at the time.
It was later changed to default to true.

Removing as it doesn't seem useful and adds confusion to the active options.
If no input is required it is probably better to use an input platform that provides this behaviour.