Add a --help-markdown flag to generate documentation
This is intended to be used by projects that are built using Mir to generate
their own documentation.

For example, if a project has a shell called `my-shell` they would generate this using:
```
$ my-shell --help-markdown > options.md.include
```

And then include that into their own documentation with:
```
Options for my-shell, which can be set on the command line, e.g. `--version`.

{include} options.md.include
```