Fix cursor size getting truncated after scaling.
Inside `mg::scale_cursor_image`, we round the width and height after
scaling as this is a precondition required by pixman. But outside, the
width and height would be _truncated_. Thus, the size we'd allocate
would sometimes be smaller than the actual buffer we get from pixman,
and this would cause buffer overruns when copying the pixman buffer into
the buffer used for display.

Example:
At scale 2.51, pixman's buffer size would be scaled and round _up_ so
the buffer could accomodate the scaled image, but the size outside would
be scaled and _truncated_ (effectively rounded down). When copying the
data over, the smaller buffer would get overrun.