Fix `miral::CursorScale(float)` not applying on launch
Fixes a bug introduced in the previous commit. We default the cursor
scale option to the scale given in the constructor. If no value is
provided to the option, or the value is the same as the one defined in
the constructor, then we attempt to set the scale to the same value
that's already stored in `Self`, which because of the previous commit
ignores the value and doesn't call `AccessibilityManager::cursor_scale`.

Solution? Bypass the check.