geometry: Add a generic Value::as<T>() accessor
This allows you to extract the value of a Value as any
type that is losslessly convertible from the underlying
type of the Value.

That is: `Value<int, Tag>::as<long>()` and
`Value<int, Tag>::as<double>()` exist, as any `int`
is representable in a `long` and any `int` is exactly
representable in a `double`, but
`Value<uint64_t, Tag>::as<int>()` and
`Value<int, Tag>::as<float>()` do not exist, as `uint64_t`
can represent values outside of the range of `int`, and
`int` can represent values not exactly representable in
`float`.