tests/wlcs: Disable the cursor image.
WLCS uses the `wl_surface.frame` callback to (attempt to) detect when a surface
is displayed. This *mostly* works, but if the cursor happens to be occluding the
surface then a new client frame is not required and so the compositor (correctly)
doesn't send a `wl_surface.frame` event.

Work around this by just using the Mir option to disable the cursor image.